requirements.txt
pyyaml
pydub
simpleaudio
pynput